requirements.txt
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
dnspython==2.3.0
idna==3.4
pycodestyle==2.10.0
pymongo==4.3.3
python-dotenv==0.21.1
requests==2.28.2
tomli==2.0.1
urllib3==1.26.14
